Minor fixes and updates
Updated plot extensions to '.jpg' to avoid transparent backgrounds that make interpretation difficult in dark mode visualization programs.
Updated some paths, and added notes.
Minor fixes on comments.
Fixed an error that prevented the data in the frustration frequency histogram (both the global and individuals) to be based on the correct data (resulting in the same plot in all cases).